examples/torcharrow/requirements.txt
click
torcharrow
torchdata
